feat: add MSVC build support.

+++ bridge/bindings/qjs/qjs_engine_patch.cc
@@ -251,7 +251,7 @@ uint16_t* JS_ToUnicode(JSContext* ctx, JSValueConst value, uint32_t* length) {
   if (!string->is_wide_char) {
     uint8_t* p = string->u.str8;
     uint32_t len = *length = string->len;
-    buffer = (uint16_t*)malloc(sizeof(uint16_t) * len * 2);
+    buffer = (uint16_t*)malloc(sizeof(uint8_t) * len * 2);
     for (size_t i = 0; i < len; i++) {
       buffer[i] = p[i];
       buffer[i + 1] = 0x00;
